- Refactoring IdManager to IdGenerator class - Changing methods, deleting possibility to sign ID manually in IdGenerator class and @Identifiable - changing default method getId() in @Identifiable - deleting redundant code
