+++ SecondTask/src/Service/TicketService.java
@@ -4,7 +4,7 @@ import Enums.StadiumSector;
 import Interfaces.Identifiable;
 import Model.Admin;
 import Model.Client;
-import Model.IdManeger;
+import Model.IdGenerator;
 import Model.Ticket;
 
 import java.time.LocalDateTime;
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 public class TicketService implements Identifiable {
     private long id;
     public static void main(String[] args) throws IllegalAccessException {
-        Ticket ticket1 = new Ticket(1L, "Bohemy", 123, LocalDateTime.now(), true, StadiumSector.A, 32);
+        Ticket ticket1 = new Ticket("Bohemy", 123, LocalDateTime.now(), true, StadiumSector.A, 32);
         System.out.println("Creation of ticket №1 took " + ticket1.getDuration() + " ms. Price: " + ticket1.getTicketPrice() + ".");
         ticket1.shared("+48 888 88 88");
 
@@ -33,18 +33,7 @@ public class TicketService implements Identifiable {
         nick.checkTicket(ticket1);
     }
 
-    @Override
-    public void setId(Long id){
-        this.id = id;
-    }
-
-    @Override
-    public Long getId(){
-        return id;
-    }
-
     public TicketService(){
-        this.id = IdManeger.generateId();
+        this.id = IdGenerator.generateId();
     }
-
 }
